refactor: the review page to be more organized and have more clear structure
