+++ apps/google-docs/src/locations/Page/components/review/mapping/buildTextSegments.ts
