+++ apps/google-docs/src/locations/Page/components/review/mapping/resolveMappingCardOffsets.ts
